+++ l10n_bg_claude_terminal/models/__init__.py
@@ -2,3 +2,9 @@
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
 
 from . import res_users
+from . import ai_view_registry
+from . import ai_composite_document
+from . import ai_view_parser
+from . import ai_embedding_provider
+from . import ai_qdrant_client
+from . import ai_document_builder